Added clock in MainView in a different way than intended
