+++ src/main/java/pl/sytomczak/nutritiontraining/gui/MenuView.form
@@ -54,11 +54,13 @@
           <text value="Stretching"/>
         </properties>
       </component>
-      <grid id="34f2f" layout-manager="GridLayoutManager" row-count="2" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
+      <grid id="34f2f" layout-manager="GridLayoutManager" row-count="1" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
         <margin top="0" left="0" bottom="0" right="0"/>
         <constraints>
           <grid row="2" column="2" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="6" fill="0" indent="0" use-parent-layout="false">
-            <maximum-size width="70" height="70"/>
+            <minimum-size width="120" height="20"/>
+            <preferred-size width="120" height="20"/>
+            <maximum-size width="120" height="20"/>
           </grid>
         </constraints>
         <properties>
@@ -69,24 +71,17 @@
           <component id="7971b" class="javax.swing.JEditorPane" binding="clockPanel">
             <constraints>
               <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="6" hsize-policy="6" anchor="6" fill="0" indent="0" use-parent-layout="false">
-                <maximum-size width="70" height="70"/>
+                <minimum-size width="120" height="20"/>
+                <preferred-size width="120" height="20"/>
+                <maximum-size width="120" height="20"/>
               </grid>
             </constraints>
             <properties>
-              <background color="-4656542"/>
+              <background color="-1576209"/>
               <editable value="false"/>
+              <font size="13" style="1"/>
             </properties>
           </component>
-          <hspacer id="4a65b">
-            <constraints>
-              <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="1" hsize-policy="6" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-            </constraints>
-          </hspacer>
-          <vspacer id="c1871">
-            <constraints>
-              <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="6" hsize-policy="1" anchor="0" fill="2" indent="0" use-parent-layout="false"/>
-            </constraints>
-          </vspacer>
         </children>
       </grid>
       <vspacer id="b218c">
